Fix tavern/ent go generate
